requirements.txt
# APEX_OMEGA_De1 · Requirements
fastapi==0.115.0
uvicorn==0.30.6
python-telegram-bot==21.5
apscheduler==3.10.4
requests==2.32.3
scipy==1.14.1
numpy==2.1.0
python-dotenv==1.0.1
httpx==0.27.2
